alterações de double para Double, criado atualização por id no service, repository e controller

+++ src/main/java/com/tads/jerseymarketapi/service/factory/ProductFactory.java
@@ -13,7 +13,7 @@ public class ProductFactory {
         ProductModel productModel = new ProductModel();
         BeanUtils.copyProperties(productDto, productModel);
         productModel.setRegistrationDate(LocalDateTime.now(ZoneId.of("America/Sao_Paulo")));
-        productModel.setRating(0);
+        productModel.setRating((double) 0);
         return productModel;
     }
 }
